+++ R/calibrate.R
@@ -336,8 +336,10 @@ mix.calibrationcurves <- function(proportion=.5, cc1="3Col_intcal20.14C", cc2="3
 
 
 # See Christen and Perez 2009, Radiocarbon 51:1047-1059. Instead of assuming the standard Gaussian model (default in clam), a student t distribution can be used with two parameters. Christen and Perez 2009 suggest t.a = 3 and t.b = 4; this can be put as clam( calibt=c(3,4) )
-.calibt <- function(t.a, t.b, f.cage, f.error, theta, f.mu, f.sigma)
+.calibt <- function(t.a, t.b, f.cage, f.error, f.mu, f.sigma) # removed theta as par
   (t.b + ((f.cage-f.mu)^2) / (2*(f.sigma^2 + f.error^2))) ^ (-1*(t.a+0.5))
+
+#(t.b + ((y-x)^2) / (2*(error^2))) ^ (-1*(t.a+0.5))
 
 #' @name student.t 
 #' @title Comparison dates calibrated using both the student-t distribution and the the normal distribution.
@@ -523,7 +525,7 @@ calBP.14C <- function(yr, cc=1, cc1="3Col_intcal20.14C", cc2="3Col_marine20.14C"
     # calibrate; find how far f.cage (measurement) is from f.mu (calibration curve)
     if(length(calibt) < 2)
       cal <- cbind(theta, dnorm(f.mu, f.cage, sqrt(f.error^2+f.sigma^2))) else
-        cal <- cbind(theta, .calibt(calibt[1], calibt[2], f.cage, f.error, theta, f.mu, f.sigma))
+        cal <- cbind(theta, .calibt(calibt[1], calibt[2], f.cage, f.error, f.mu, f.sigma))
 
     # interpolate and normalise calibrated distribution to 1
     cal <- cal[min(which(cal[,2] > 0)):max(which(cal[,2] > 0)),] # remove unnecessary data